refactor: split repository-list into a dumb component
