+++ src/screens/organization/repository-list.js
@@ -1,45 +1,14 @@
 /* eslint-disable no-shadow */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { FlatList, View, Dimensions, StyleSheet } from 'react-native';
 
-import {
-  ViewContainer,
-  RepoListItem,
-  LoadingRepositoryListItem,
-  SearchBar,
-} from 'components';
-import { colors } from 'config';
+import { RepositoryList } from 'components';
 
 import { getRepos } from 'api/rest/providers/github/endpoints/orgs';
 import { searchRepos } from 'api/rest/providers/github/endpoints/search';
 
-const loadData = ({ orgId, getRepos }) => {
-  getRepos(orgId);
-};
-
-// TODO: clean up searchStart & query usage
-
-const styles = StyleSheet.create({
-  header: {
-    borderBottomColor: colors.greyLight,
-    borderBottomWidth: 1,
-  },
-  searchBarWrapper: {
-    flexDirection: 'row',
-  },
-  searchContainer: {
-    width: Dimensions.get('window').width,
-    backgroundColor: colors.white,
-    flex: 1,
-  },
-  searchCancelButton: {
-    color: colors.black,
-  },
-  listContainer: {
-    marginBottom: 90,
-  },
-});
+const getQueryString = (keyword, orgId) =>
+  `q=${keyword}+user:${orgId}+fork:true&per_page=8`;
 
 class OrgRepositoryList extends Component {
   props: {
@@ -58,132 +27,34 @@ class OrgRepositoryList extends Component {
     navigation: Object,
   };
 
-  state: {
-    query: string,
-    searchStart: boolean,
-    searchFocus: boolean,
-  };
-
-  constructor() {
-    super();
-
-    this.state = {
-      query: '',
-      searchStart: false,
-      searchFocus: false,
-    };
-
-    this.search = this.search.bind(this);
-    this.getList = this.getList.bind(this);
-  }
-
-  componentWillMount() {
-    loadData(this.props);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.orgId !== this.props.orgId) {
-      loadData(nextProps);
-    }
-  }
-
-  getList = () => {
-    const { searchedResults, repositories } = this.props;
-    const { searchStart } = this.state;
-
-    return searchStart ? searchedResults : repositories;
-  };
-
-  loadMore = () => {
-    const { orgId, getRepos, searchRepos, navigation } = this.props;
-
-    if (navigation.state.params.searchedKeyword) {
-      searchRepos(navigation.state.params.searchedKeyword, true);
-    } else {
-      getRepos(orgId, { loadMore: true });
-    }
-  };
-
-  search(query) {
-    if (query !== '') {
-      const searchedKeyword = `q=${query}+user:${this.props.navigation.state
-        .params.orgId}+fork:true`;
-
-      this.setState({
-        searchStart: true,
-        query,
-      });
-
-      const { searchRepos, navigation } = this.props;
-
-      navigation.setParams({ searchedKeyword });
-      searchRepos(searchedKeyword);
-    }
-  }
-
-  keyExtractor = item => {
-    return item.id;
-  };
-
   render() {
     const {
+      orgId,
       authUser,
       navigation,
-      searchedResultsPagination,
+      getRepos,
+      searchRepos,
+      repositories,
       repositoriesPagination,
+      searchedResults,
+      searchedResultsPagination,
     } = this.props;
-    const repoCount = navigation.state.params.repoCount;
-    const { searchStart, searchFocus } = this.state;
-    const loading =
-      (searchedResultsPagination.isFetching &&
-        !searchedResultsPagination.pageCount) ||
-      (repositoriesPagination.isFetching && !repositoriesPagination.pageCount);
 
     return (
-      <ViewContainer>
-        <View>
-          <View style={styles.header}>
-            <View style={styles.searchBarWrapper}>
-              <View style={styles.searchContainer}>
-                <SearchBar
-                  textColor={colors.primaryDark}
-                  textFieldBackgroundColor={colors.greyLight}
-                  showsCancelButton={searchFocus}
-                  onFocus={() => this.setState({ searchFocus: true })}
-                  onCancelButtonPress={() =>
-                    this.setState({ searchStart: false, query: '' })}
-                  onSearchButtonPress={query => {
-                    this.search(query);
-                  }}
-                  hideBackground
-                />
-              </View>
-            </View>
-          </View>
+      <RepositoryList
+        authUser={authUser}
+        navigation={navigation}
+        loadRepositories={(loadMore = false) => getRepos(orgId, { loadMore })}
+        loadSearchResults={(keyword, loadMore = false) => {
+          navigation.setParams({ searchedKeyword: keyword });
 
-          {loading &&
-            [...Array(searchStart ? repoCount : 10)].map(
-              (item, index) => <LoadingRepositoryListItem key={index} /> // eslint-disable-line react/no-array-index-key
-            )}
-
-          {!loading &&
-            <View style={styles.listContainer}>
-              <FlatList
-                removeClippedSubviews={false}
-                data={this.getList()}
-                keyExtractor={this.keyExtractor}
-                onEndReached={this.loadMore}
-                onEndReachedThreshold={0.4}
-                renderItem={({ item }) =>
-                  <RepoListItem
-                    repository={item}
-                    showFullName={true || authUser.login !== item.owner.login}
-                    navigation={navigation}
-                  />}
-              />
-            </View>}
-        </View>
-      </ViewContainer>
+          return searchRepos(getQueryString(keyword, orgId), { loadMore });
+        }}
+        repositories={repositories}
+        repositoriesPagination={repositoriesPagination}
+        searchResults={searchedResults}
+        searchResultsPagination={searchedResultsPagination}
+      />
     );
   }
 }
@@ -201,9 +72,12 @@ const mapStateToProps = (state, ownProps) => {
   const repositories = repositoriesPagination.ids.map(id => repos[id]);
 
   const searchedKeyword = ownProps.navigation.state.params.searchedKeyword;
-  const searchedResultsPagination = searchedKeyword
-    ? reposBySearch[searchedKeyword]
-    : { ids: [] };
+  const queryString = getQueryString(searchedKeyword, orgId);
+
+  const searchedResultsPagination =
+    searchedKeyword && reposBySearch[queryString]
+      ? reposBySearch[queryString]
+      : { ids: [] };
   const searchedResults = searchedResultsPagination.ids.map(id => repos[id]);
 
   return {
